dziennik/urls.py
from django.contrib import admin
from django.urls import path, include
from django.contrib.auth import views as auth_views
from allauth.account import views as allauth_views
from . import views

urlpatterns = [
    path('', views.main, name='main'),
    #path('', include('django.contrib.auth.urls')),
    path('panel/admin/', admin.site.urls),
    path('accounts/', include('allauth.urls')),
    #path('login/',auth_views.LoginView.as_view(template_name='login/login.html'), name='login'),
    path('login/',allauth_views.LoginView.as_view(template_name='login/login.html'), name='login'),
    path('logout/',auth_views.LogoutView.as_view(template_name='base.html'), name='logout'),
    path('singup/', views.singup, name='singup'),
]
