requirement.txt
ansicon==1.89.0
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.4
django-admin-logs==1.0.1
django-allauth==0.44.0
django-phone-field==1.8.1
django-picklefield==3.0.1
idna==2.10
jinxed==1.0.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wcwidth==0.2.5
